Corrected code via advices

+++ src/main/java/com/example/bank/model/entities/Transaction.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -20,7 +21,7 @@ public class Transaction {
     private long id;
 
     @Column(name = "value")
-    private long value;
+    private BigDecimal value;
 
     @ManyToOne
     @JoinColumn(name = "card_id_to")
